Fix initial image position (closes #13)
It seems the problem is caused because the window is resized.
For example, when there's a filter the height needs to be increased to
make room for the filter parameters. Because of the timing of actions,
this resizing happens after the image has been displayed, but it does
not trigger a resizing/repositioning of the ScrolledWindow with the
image.